requirements.txt
Django>=1.5
ez-setup>=0.9
django-openid-auth>=0.4
django-dajax>=0.9
django-dajaxice>=0.5
Markdown>=2.2
pygments>=1.4
-e git://github.com/9h37/django-gitstorage.git#egg=django-gitstorage
python-openid>=2.2.5
simplejson>=2.6.2
